+++ Sources/Sentry/include/SentrySDK.h
@@ -31,20 +31,6 @@ SENTRY_NO_INIT
 + (void)setCurrentHub:(SentryHub *)hub;
 
 /**
- * Use [SentrySDK startWithOptionsObject] / SentrySDK.start(:) instead
- * @deprecated
- */
-+ (instancetype)initWithOptionsObject:(SentryOptions *)options NS_SWIFT_NAME(init(options:))
-                                          __attribute((deprecated(("Use startWithOptionsObject"))));
-
-/**
- * Use [SentrySDK startWithOptions] / SentrySDK.start(:) instead
- * @deprecated
- */
-+ (instancetype)initWithOptions:(NSDictionary<NSString *, id> *)optionsDict
-    NS_SWIFT_NAME(init(options:))__attribute((deprecated(("Use startWithOptions"))));
-
-/**
  * Inits and configures Sentry (SentryHub, SentryClient) and sets up all integrations.
  */
 + (void)startWithOptions:(NSDictionary<NSString *, id> *)optionsDict NS_SWIFT_NAME(start(options:));