+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - fix: Umbrella header #671
+- feat: Remove deprecated SDK inits #673
 - fix: Crash in SentryEnvelope.initWithEvent #643
 - fix: Build failure for Apple Silicon Macs #588